create main_test
