+++ pkg/broker/configurator_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/Goboolean/shared-packages/pkg/broker"
 	"github.com/Goboolean/shared-packages/pkg/resolver"
-	"github.com/joho/godotenv"
 )
 
 var (
@@ -18,20 +17,7 @@ var (
 
 
 
-func TestMain(m *testing.M) {
 
-	if err := os.Chdir("../../"); err != nil {
-		panic(err)
-	}
-
-	if err := godotenv.Load(); err != nil {
-		panic(err)
-	}
-
-	code := m.Run()
-
-	os.Exit(code)
-}
